Fix Cameras not Working Together
