+++ src/components/KillCam.tsx
@@ -6,24 +6,23 @@ import {
   matchDimensions,
   detectAllFaces,
   resizeResults,
-  draw
+  draw,
+  loadSsdMobilenetv1Model,
+  loadFaceLandmarkModel,
+  loadFaceRecognitionModel,
+  LabeledFaceDescriptors
+
 } from 'face-api.js';
 import { useWebcam } from '../contexts/WebcamProvider';
 import TargetRecognition from './KillProgress';
 import { useAuth0 } from "@auth0/auth0-react";
-
 import SocketContext from '../contexts/Socket/SocketContext';
 
 
-type KillCamProps = {
-  faceMatcher: (FaceMatcher | null)
-}
-
-const KillCam: React.FC<KillCamProps> = ({ faceMatcher }) => {
+const KillCam: React.FC = () => {
   const webcamContext = useWebcam();
   const webcamRef = webcamContext?.webcamRef;
   const videoStarted = webcamContext?.videoStarted;
-  const navigate = useNavigate();
   const videoHeight = window.innerHeight;
   const videoWidth = window.innerWidth;
   const displaySize = { width: videoWidth, height: videoHeight };
@@ -32,6 +31,8 @@ const KillCam: React.FC<KillCamProps> = ({ faceMatcher }) => {
   let wasBountyDetected = false;
   const { user } = useAuth0();
 
+  const [faceMatcher, setFaceMatcher] = useState<FaceMatcher | null>(null);
+
   // socket contexts
   const { games, users } = useContext(SocketContext).SocketState;
   const { UpdateGameStatus, AddGameStats } = useContext(SocketContext);
@@ -39,19 +40,42 @@ const KillCam: React.FC<KillCamProps> = ({ faceMatcher }) => {
   // storing username of hunted so tensor can compare
   const [huntedUsername, setHuntedUsername] = useState<string>('');
 
+  useEffect(() => {
+    if (users.length > 0) {
+      loadTensorFlowFaceMatcher();
+    }
+  }, [users]);
+
+  const loadTensorFlowFaceMatcher = async () => {
+    try {
+      await loadSsdMobilenetv1Model('/models')
+      await loadFaceLandmarkModel('/models')
+      await loadFaceRecognitionModel('/models')
+      await createFaceMatcher();
+    } catch (err) {
+      console.error(err);
+    }
+  };
+
+  const createFaceMatcher = async () => {
+    const labeledFaceDescriptors = users.map((player) => {
+      // Convert each user's description array back to a Float32Array
+      const descriptions = [new Float32Array(player.facialDescriptions)];
+      return new LabeledFaceDescriptors(player.username, descriptions);
+    });
+    setFaceMatcher(new FaceMatcher(labeledFaceDescriptors, 0.5));
+  }
+
+  const createCanvas = () => {
+    if (videoStarted && webcamRef?.current?.video) {
+      canvasRef.current = createCanvasFromMedia(webcamRef.current.video);
+    }
+  };
 
   // creating the canvas when the component mounts
   useEffect(() => {
     createCanvas();
-    return () => {
-      // Stop the camera here using webcamRef.current
 
-      // Example:
-      if (webcamRef?.current?.video) {
-        webcamRef.current.video.pause();
-        webcamRef.current.video.srcObject = null;
-      }
-    };
   }, [])
 
   // if faceMatcher and huntedUsername are both set, start the handleVideoPlay()
@@ -81,11 +105,6 @@ const KillCam: React.FC<KillCamProps> = ({ faceMatcher }) => {
     }
   }, [users])
 
-  const createCanvas = () => {
-    if (videoStarted && webcamRef?.current?.video) {
-      canvasRef.current = createCanvasFromMedia(webcamRef.current.video);
-    }
-  };
 
   const updateCounter = () => {
     if (wasBountyDetected) {
@@ -98,37 +117,41 @@ const KillCam: React.FC<KillCamProps> = ({ faceMatcher }) => {
   };
 
   const handleVideoOnPlay = () => {
-    if (canvasRef.current && webcamRef?.current) {
-      matchDimensions(canvasRef.current, displaySize);
-      const context = canvasRef.current.getContext('2d');
-      if (context) {
-        const processVideoFrame = async () => {
-          if (faceMatcher && webcamRef?.current?.video && canvasRef.current) {
-            const detections = await detectAllFaces(webcamRef.current.video)
-              .withFaceLandmarks()
-              .withFaceDescriptors();
-            const resizedDetections = resizeResults(detections, displaySize);
-            context.clearRect(0, 0, videoWidth, videoHeight);
-            draw.drawFaceLandmarks(canvasRef.current, resizedDetections);
-            const results = resizedDetections.map(d => faceMatcher.findBestMatch(d.descriptor))
-            results.forEach((result, i) => {
-              const name = result.toString()
-              const sliceIndex = name.indexOf(' (')
-              const detectedFace = name.slice(0, sliceIndex)
-              if (detectedFace === huntedUsername) {
-                wasBountyDetected = true;
-              }
-              const box = resizedDetections[i].detection.box
+    if (!canvasRef.current || !webcamRef?.current) return;
+    matchDimensions(canvasRef.current, displaySize);
+    const context = canvasRef.current.getContext('2d');
+    if (context) {
+      const processVideoFrame = async () => {
+        if (faceMatcher && webcamRef?.current?.video && canvasRef.current) {
+          const detections = await detectAllFaces(webcamRef.current.video)
+            .withFaceLandmarks()
+            .withFaceDescriptors();
+          const resizedDetections = resizeResults(detections, displaySize);
+          context.clearRect(0, 0, videoWidth, videoHeight);
+          draw.drawFaceLandmarks(canvasRef.current, resizedDetections);
+          const results = resizedDetections.map(d => faceMatcher.findBestMatch(d.descriptor))
+          results.forEach((result, i) => {
+            const name = result.toString()
+            const sliceIndex = name.indexOf(' (')
+            const detectedFace = name.slice(0, sliceIndex)
+            if (detectedFace === huntedUsername) {
+              wasBountyDetected = true;
+            }
+            const box = resizedDetections[i].detection.box
+            if (box) {
               const drawBox = new draw.DrawBox(box, { label: name })
               if (canvasRef.current) {
                 drawBox.draw(canvasRef.current)
               }
-            });
-            setTimeout(processVideoFrame, 100);
-          }
-        };
-        setTimeout(processVideoFrame, 100);
-      }
+            }
+
+          });
+          requestAnimationFrame(processVideoFrame);
+        }
+      };
+      requestAnimationFrame(processVideoFrame);
+    } else {
+      return;
     }
   };
